# api.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Cards API
paths:
  /cards:
    get:
      operationId: listCards
      parameters:
        - name: limit
          in: query
          description: maximum number of results to return
          required: false
          schema:
            type: integer
            format: int32
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: addCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCard'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cards/{id}:
    get:
      operationId: findCardByID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Card:
      allOf:
        - $ref: '#/components/schemas/NewCard'
        - required:
            - id
          properties:
            id:
              type: string

    NewCard:
      required:
        - owner
      properties:
        owner:
          type: string
          description: Owner of the card

    Error:
      required:
        - code
      properties:
        code:
          type: integer
        message:
          type: string