# api/models.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package api

// Card defines model for Card.
type Card struct {
	Id string `json:"id"`

	// Owner Owner of the card
	Owner string `json:"owner"`
}

// Error defines model for Error.
type Error struct {
	Code    int     `json:"code"`
	Message *string `json:"message,omitempty"`
}

// NewCard defines model for NewCard.
type NewCard struct {
	// Owner Owner of the card
	Owner string `json:"owner"`
}

// ListCardsParams defines parameters for ListCards.
type ListCardsParams struct {
	// Limit maximum number of results to return
	Limit *int32 `form:"limit,omitempty" json:"limit,omitempty"`
}

// AddCardJSONRequestBody defines body for AddCard for application/json ContentType.
type AddCardJSONRequestBody = NewCard